Organise modules and configuration consistently
