+++ package.json
@@ -26,10 +26,10 @@
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "typeorm": "typeorm-ts-node-esm",
     "migration:create": "typeorm-ts-node-esm migration:create",
-    "migration:run": "typeorm-ts-node-esm -d ./src/config/typeorm.datasource.ts migration:run",
-    "migration:revert": "typeorm-ts-node-esm -d ./src/config/typeorm.datasource.ts migration:revert",
-    "seed:run": "typeorm-ts-node-esm -d ./src/config/typeorm-seeds.datasource.ts migration:run",
-    "seed:revert": "typeorm-ts-node-esm -d ./src/config/typeorm-seeds.datasource.ts migration:revert",
+    "migration:run": "typeorm-ts-node-esm -d ./src/database/typeorm.datasource.ts migration:run",
+    "migration:revert": "typeorm-ts-node-esm -d ./src/database/typeorm.datasource.ts migration:revert",
+    "seed:run": "typeorm-ts-node-esm -d ./src/database/typeorm-seeds.datasource.ts migration:run",
+    "seed:revert": "typeorm-ts-node-esm -d ./src/database/typeorm-seeds.datasource.ts migration:revert",
     "firebase:token": "node ./scripts/get-token.js",
     "deploy:production": "pm2 deploy ecosystem.config.json production update",
     "deploy:staging": "pm2 deploy ecosystem.config.json staging update",
