+++ src/cli.ts
@@ -1,7 +1,7 @@
 import { NestFactory } from '@nestjs/core'
 import { CommandModule, CommandService } from 'nestjs-command'
-import { AppModule } from './app.module'
-;(async () => {
+import { AppModule } from './app'
+void (async () => {
   const app = await NestFactory.createApplicationContext(AppModule, {
     logger: ['log', 'error', 'warn', 'debug', 'verbose'],
   })
