+++ src/config/index.ts
@@ -1,9 +1,3 @@
-export * from './body-size.config'
-export * from './cors.config'
+export * from './config.module'
 export * from './graceful-shutdown.config'
-export * from './json.middleware'
-export * from './recaptcha.config'
-export * from './schema.config'
-export * from './swagger.config'
-export * from './typeorm.config'
 export * from './validator-container.config'
