+++ src/database/database.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common'
+import { ConfigService } from '@nestjs/config'
+import { TypeOrmModule } from '@nestjs/typeorm'
+import { ConfigModule } from '../config'
+import { TypeOrmConfigService } from './typeorm.config'
+
+@Module({
+  imports: [
+    TypeOrmModule.forRootAsync({
+      imports: [ConfigModule],
+      inject: [ConfigService],
+      useClass: TypeOrmConfigService,
+    }),
+  ],
+})
+export class DatabaseModule {}
