--- src/filters/index.ts
@@ -1,2 +1,0 @@
-export * from './not-found-exception.filter'
-export * from './i18n-exceptions.filter'
