--- src/firebase/firebase-admin-sdk.type.ts
@@ -1,3 +1,0 @@
-import * as admin from 'firebase-admin'
-
-export type FirebaseAdminSDK = admin.app.App
