+++ src/firebase/firebase-admin.constants.ts
@@ -1,3 +1,2 @@
 export const FIREBASE_ADMIN_MODULE_OPTIONS = 'FIREBASE_ADMIN_MODULE_OPTIONS'
-export const FIREBASE_ADMIN_NAME = 'FIREBASE_ADMIN_NAME'
 export const FIREBASE_ADMIN_INJECT = 'FIREBASE_ADMIN_INJECT'
