+++ src/firebase/index.ts
@@ -1,5 +1,2 @@
 export * from './auth'
-export * from './firebase-admin-sdk.type'
-export * from './firebase-admin.constant'
-export * from './firebase-admin.interface'
-export * from './firebase-admin.module'
+export * from './firebase.module'
