+++ src/http/body-size.config.ts
@@ -1,5 +1,5 @@
 import { INestApplication } from '@nestjs/common'
-import { jsonMiddleware } from '.'
+import { jsonMiddleware } from './json.middleware'
 
 export function setBodySize(app: INestApplication): void {
   app.use(jsonMiddleware)
