+++ src/protocols/api/protocol-assignees.controller.ts
@@ -20,7 +20,7 @@ import { UsersRepository } from '../..//users/entities/users.repository'
 import {
   AcceptedResponse,
   ACCEPTED_RESPONSE_STATUS,
-} from '../../utils/accepted-response'
+} from '../../http/accepted-response'
 import { InjectUser } from '../../auth/decorators/inject-user.decorator'
 import { User } from '../../users/entities'
 import { Protocol } from '../entities/protocol.entity'
