+++ src/security/security.module.ts
@@ -1,0 +1,30 @@
+import { Module } from '@nestjs/common'
+import { ConfigModule, ConfigService } from '@nestjs/config'
+import { GoogleRecaptchaModule } from '@nestlab/google-recaptcha'
+import { GoogleRecaptchaConfigService } from './recaptcha.config'
+import { ThrottlerModule, ThrottlerGuard } from '@nestjs/throttler'
+import { ThrottlerConfig } from './throttler.config'
+import { APP_GUARD } from '@nestjs/core'
+
+@Module({
+  imports: [
+    ConfigModule,
+    ThrottlerModule.forRootAsync({
+      inject: [ConfigService],
+      imports: [ConfigModule],
+      useClass: ThrottlerConfig,
+    }),
+    GoogleRecaptchaModule.forRootAsync({
+      imports: [ConfigModule],
+      inject: [ConfigService],
+      useClass: GoogleRecaptchaConfigService,
+    }),
+  ],
+  providers: [
+    {
+      provide: APP_GUARD,
+      useClass: ThrottlerGuard,
+    },
+  ],
+})
+export class SecurityModule {}
