+++ test/app.e2e-spec.ts
@@ -1,6 +1,6 @@
 import * as request from 'supertest'
 import { Test } from '@nestjs/testing'
-import { AppModule } from '../src/app.module'
+import { AppModule } from '../src/app'
 import { INestApplication } from '@nestjs/common'
 
 describe('AppController (e2e)', () => {
